Separate out spec inspection implementation from activation-factory aspect of AgentToolRegistry
